+++ module/src/helpers.ps1
@@ -7,7 +7,7 @@ function Write-tftoolsLogo {
     Write-Host " / __/ /_/ __/ __ \/ __ \/ / ___/" -ForegroundColor DarkMagenta
     Write-Host "/ /_/ __/ /_/ /_/ / /_/ / (__  ) " -ForegroundColor DarkMagenta
     Write-Host "\__/_/  \__/\____/\____/_/____/  " -ForegroundColor DarkMagenta
-    Write-Host "                      v0.3.1     " -ForegroundColor DarkGray
+    Write-Host "                      v0.3.5     " -ForegroundColor DarkGray
 }
 # This one is pretty cool, and we could probably have used a different
 # module for working with Zip files but this is a cool scripting exercise
@@ -33,9 +33,17 @@ function Set-PlatformVariables {
         $global:machineOS = "windows_amd64"
         $global:execDir = "$env:LOCALAPPDATA\Microsoft\WindowsApps"
     }
-    elseif ($PSVersionTable.Platform -eq "Unix") {
+    elseif ($PSVersionTable.OS -like "Linux*") {
         $global:tfPath = $HOME + "/.tftools"
         $global:machineOS = "linux_amd64"
+        if ($env:PATH -notlike "*$tfPath*") {
+            $env:PATH += $tfPath
+            '$env:PATH' + " += $tfPath" | Add-Content -Path $PROFILE
+        }
+    }
+    elseif ($PSVersionTable.OS -like "Darwin*") {
+        $global:tfPath = $HOME + "/.tftools"
+        $global:machineOS = "darwin_amd64"
         if ($env:PATH -notlike "*$tfPath*") {
             $env:PATH += $tfPath
             '$env:PATH' + " += $tfPath" | Add-Content -Path $PROFILE